pyproject.toml
[project]
name = "civicaurai"
version = "0.1.0"
description = "CivicAurAI Semantic Spanner Engine"
authors = [
    { name = "Csaba Toth", email = "csaba.toth.us@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-cloud-aiplatform>=1.60.0",
    "google-cloud-spanner>=3.40.0",
    "google-cloud-storage>=2.14.0",
    "opencv-python>=4.9.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "numpy>=1.26",
    "Pillow>=10.0"
]
